Move worker test helpers into base junebug test base so we can use them for channel status worker tests
